+++ gqlalchemy/vendors/memgraph.py
@@ -17,16 +17,16 @@ import sqlite3
 
 from typing import List, Optional, Union
 
-from .connection import Connection, MemgraphConnection
+from ..connection import Connection, MemgraphConnection
 from .database import Database
-from .disk_storage import OnDiskPropertyDatabase
-from .graph_algorithms.query_modules import QueryModule
-from .exceptions import (
+from ..disk_storage import OnDiskPropertyDatabase
+from ..graph_algorithms.query_modules import QueryModule
+from ..exceptions import (
     GQLAlchemyError,
     GQLAlchemyUniquenessConstraintError,
     GQLAlchemyOnDiskPropertyDatabaseNotDefinedError,
 )
-from .models import (
+from ..models import (
     MemgraphConstraintExists,
     MemgraphConstraintUnique,
     MemgraphIndex,
